+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.0.0] - 2026-05-22
+
+### Added
+
+- iOS/macOS can initialize the SDK using a PList file
+- Android can initialize SDK using a JSON file
+
+### Changed
+
+- Updated iOS/macOS native SDK to 5.0.0
+- Updated Android native SDK to 5.0.0
+
 ## [4.3.0] - 2026-05-14
 
 ### Added